# Remove copying of range tombstones keys in iterator